[HUMAN App] refactor: my jobs table components (#3186)
